requirements.txt
requests==2.32.3
pathlib==1.0.1
python-telegram-bot==13.0
python-dotenv==1.1.0